shifted layout var and func to chatlayout
