+++ packages/react/src/hooks/useCustomTheme.js
@@ -6,7 +6,7 @@ const invertMode = (mode) => (mode === 'light' ? 'dark' : 'light');
 export const useCustomTheme = () => {
   const { ECOptions } = useRCContext();
   const theme = useTheme();
-  const { mode } = ECOptions;
+  const mode = ECOptions?.mode;
   const colors = theme.schemes[mode];
   const invertedColors = theme.schemes[invertMode(mode)];
